move item-item into knn package
