+++ docs/knn.rst
@@ -1,6 +1,8 @@
 k-NN Collaborative Filtering
 ============================
 
+.. module:: lenskit.algorithms.knn
+
 LKPY provides user- and item-based classical k-NN collaborative Filtering
 implementations.  These lightly-configurable implementations are intended
 to capture the behavior of the Java-based LensKit implementations to provide
@@ -34,7 +36,7 @@ rating values) with additional parameters to the constructor.
 Item-based k-NN
 ---------------
 
-.. module:: lenskit.algorithms.item_knn
+.. module:: lenskit.algorithms.knn.item
 
 This is LensKit's item-based k-NN model, based on the description by :cite:t:`Deshpande2004-ht`.
