+++ lenskit/algorithms/knn/__init__.py
@@ -1,0 +1,3 @@
+"""
+k-NN recommender models.
+"""
