+++ utils/recommend.py
@@ -34,7 +34,7 @@ from docopt import docopt
 
 from lenskit import batch
 from lenskit.algorithms import Recommender
-from lenskit.algorithms.item_knn import ItemItem
+from lenskit.algorithms.knn.item import ItemItem
 from lenskit.datasets import MovieLens
 
 _log = logging.getLogger("test-algo")
