+++ utils/train-model.py
@@ -25,7 +25,7 @@ import sys
 from docopt import docopt
 
 from lenskit.algorithms import Recommender
-from lenskit.algorithms.item_knn import ItemItem
+from lenskit.algorithms.knn.item import ItemItem
 from lenskit.datasets import MovieLens
 
 _log = logging.getLogger("train-model")
